skatlaz_avadratv/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Uploads==0.2.1
Pillow==10.0.0
python-magic==0.4.27
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
moviepy==1.0.3
mutagen==1.46.0
requests==2.31.0
